requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
requests>=2.31.0
openai-whisper>=20240930
torch>=2.2.0
transformers>=4.37.0
gTTS>=2.5.0
pydub>=0.25.1
numpy>=1.26.0
sentencepiece>=0.1.99
sacremoses>=0.1.1